+++ sgi-csp-service/src/test/resources/scripts/linea_investigacion.sql
@@ -9,3 +9,9 @@ INSERT INTO test.linea_investigacion_nombre (linea_investigacion_id, lang, value
 (1, 'es', 'Psicología Laboral u Organizacional'),
 (2, 'es', 'Psicología Clínica de la Salud'),
 (3, 'es', 'Psicología general');
+
+-- LINEA_INVESTIGACION_DESCRIPCION
+INSERT INTO test.linea_investigacion_descripcion (linea_investigacion_id, lang, value_) VALUES
+(1, 'es', 'Descripción de la línea de investigación 1'),
+(2, 'es', 'Descripción de la línea de investigación 2'),
+(3, 'es', 'Descripción de la línea de investigación 3');