+++ sgi-webapp/src/app/core/services/csp/linea-investigacion/linea-investigacion-response.ts
@@ -1,7 +1,8 @@
-import { I18nFieldValueResponse } from "@core/i18n/i18n-field-response";
+import { I18nFieldValueResponse } from '@core/i18n/i18n-field-response';
 
 export interface ILineaInvestigacionResponse {
   id: number;
   nombre: I18nFieldValueResponse[];
+  descripcion: I18nFieldValueResponse[];
   activo: boolean;
 }